fix: Fix for collection config not found
The collection config package that is assembled by the endorser removes
configs for collections in which there was no write set and therefore
the error 'config for collection [xxx] not found' results. Thes commit
removes the dependency on the collection config package and uses the
collection config retriever to manage collection configs.

closes #98

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>