chore: Use unique registration IDs for protos
The current version of fabric-mod still uses the built-in protos, preventing integration with the latest version of fabric-sdk-go which uses fabric-protos-go (due to duplicate registration errors). As an interim step, this patch registers the protos under a unique name in order to prevent the duplicate errors. Once fabric-mod is updated with the latest fabric, all of this code will bew deleted.

closes #123

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>