feat: Register a query executor provider along with state DB
When a channel is created then, along with registering a state database, a query executor provider is also registered. The query executor provider returns query executors with and without a ledger lock. In some cases (such as with off-ledger and transient data) a ledger lock is not required and this will improve performance when reading data.

Also updated integration test check for required images.

closes #291

Signed-off-by: Bob Stasyszyn <Bob.Stasyszyn@securekey.com>